# Conflicts:
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/controller/LecturerController.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/controller/UserController.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/UserRepository.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/MeetingRequestService.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/impl/EmptySlotServiceImpl.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/impl/MeetingRequestServiceImpl.java
#	Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/impl/UserServiceImpl.java

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/MeetingRequestRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MeetingRequestRepository extends JpaRepository<MeetingRequest, Integer> {
+public interface MeetingRequestRepository extends JpaRepository<MeetingRequest, Long> {
 
 }
